apps/assistant/requirements.txt
requests
python-decouple
psycopg2
ollama
opencv-python
ultralytics
kokoro
openWakeWord
faster-whisper
numpy
pyaudio
wave
